Feature/add payment/v.0.0.1

+++ Modules/Product/Http/Controllers/ProductController.php
@@ -25,7 +25,7 @@ class ProductController extends Controller
      */
     public function index()
     {
-        // abort_if(Gate::denies('access_products'), 403);
+        abort_if(Gate::denies('show_products'), 403);
         $products = ProductResource::collection(Product::with('category')->where('user_id', auth()->user()->id)->paginate(10));
         $categories = CategoryResource::collection(Category::where('user_id', auth()->user()->id)->get());
         return Inertia::render('Products/Index', ['products' => $products,'categories' => $categories]);
@@ -38,7 +38,7 @@ class ProductController extends Controller
      */
     public function create()
     {
-        // abort_if(Gate::denies('create_products'), 403);
+        abort_if(Gate::denies('create_products'), 403);
         
         return Inertia::render('Product/AddForm', [
             'urlPost' => route('product.create.store'),
@@ -54,8 +54,8 @@ class ProductController extends Controller
      */
     public function store(ProductRequest $request, Product $product)
     {
-        // abort_if(Gate::denies('show_products'), 403);
-        // dd($request->all());
+        abort_if(Gate::denies('create_products'), 403);
+
         $product->fill($request->only($product->getFillable()));
         $product->user_id = auth()->user()->id;
         $product->save();
@@ -85,7 +85,7 @@ class ProductController extends Controller
      */
     public function edit(Product $product)
     {
-        // abort_if(Gate::denies('edit_products'), 403);
+        abort_if(Gate::denies('edit_products'), 403);
 
         return Inertia::render('Product/EditForm', [
             'urlPost' => route('product.edit.post', $product->id),
@@ -103,6 +103,8 @@ class ProductController extends Controller
      */
     public function update(ProductRequest $request, Product $product)
     {
+        abort_if(Gate::denies('edit_products'), 403);
+
         $product->fill($request->only($product->getFillable()));
         if($product->isDirty()) $product->save();
 
@@ -121,7 +123,7 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
-        // abort_if(Gate::denies('delete_products'), 403);
+        abort_if(Gate::denies('delete_products'), 403);
         
         $product->delete();
         return redirect()->route('product.index');
